requirements.txt
apache-airflow-providers-amazon==8.16.0